+++ src/cascader/cascader.component.ts
@@ -3,13 +3,14 @@ import {
     InputBoolean,
     InputNumber,
     ScrollToService,
-    TabIndexDisabledControlValueAccessorMixin
+    TabIndexDisabledControlValueAccessorMixin,
+    ThyClickDispatcher
 } from 'ngx-tethys/core';
 import { ThyEmptyComponent } from 'ngx-tethys/empty';
 import { ThyIconComponent } from 'ngx-tethys/icon';
 import { SelectControlSize, SelectOptionBase, ThySelectControlComponent } from 'ngx-tethys/shared';
 import { Id } from 'ngx-tethys/types';
-import { coerceBooleanProperty, elementMatchClosest, helpers, isArray, isEmpty, set } from 'ngx-tethys/util';
+import { coerceBooleanProperty, elementMatchClosest, isArray, isEmpty, set, helpers } from 'ngx-tethys/util';
 import { BehaviorSubject, Subject } from 'rxjs';
 import { debounceTime, distinctUntilChanged, filter, take, takeUntil } from 'rxjs/operators';
 
@@ -21,7 +22,7 @@ import {
     ConnectionPositionPair,
     ViewportRuler
 } from '@angular/cdk/overlay';
-import { NgClass, NgFor, NgIf, NgStyle, NgTemplateOutlet } from '@angular/common';
+import { NgClass, NgFor, NgIf, NgStyle, NgTemplateOutlet, isPlatformBrowser } from '@angular/common';
 import {
     ChangeDetectorRef,
     Component,
@@ -29,10 +30,13 @@ import {
     EventEmitter,
     forwardRef,
     HostListener,
+    Inject,
     Input,
+    NgZone,
     OnDestroy,
     OnInit,
     Output,
+    PLATFORM_ID,
     QueryList,
     TemplateRef,
     ViewChild,
@@ -44,6 +48,7 @@ import { useHostRenderer } from '@tethys/cdk/dom';
 import { ThyCascaderOptionComponent } from './cascader-li.component';
 import { ThyCascaderSearchOptionComponent } from './cascader-search-option.component';
 import { ThyCascaderExpandTrigger, ThyCascaderOption, ThyCascaderSearchOption, ThyCascaderTriggerType } from './types';
+import { deepCopy } from '@angular-devkit/core';
 
 function toArray<T>(value: T | T[]): T[] {
     let ret: T[];
@@ -293,6 +298,12 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
     @Input() @InputBoolean() thyShowSearch: boolean = false;
 
     /**
+     * 多选选中项的展示方式，默认为空，渲染文字模板，传入tag，渲染展示模板,
+     * @default ''｜tag
+     */
+    @Input() thyPreset: string = '';
+
+    /**
      * 值发生变化时触发，返回选择项的值
      * @type EventEmitter<any[]>
      */
@@ -380,7 +391,7 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
 
     private isMultiple = false;
 
-    private prevSelectedOptions: ThyCascaderOption[] = [];
+    private prevSelectedOptions: Set<ThyCascaderOption> = new Set<ThyCascaderOption>();
 
     public menuMinWidth = 122;
 
@@ -402,6 +413,8 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
 
     private leafNodes: ThyCascaderSearchOption[] = [];
 
+    private valueChange$ = new Subject();
+
     ngOnInit(): void {
         this.setClassMap();
         this.setMenuClass();
@@ -421,6 +434,28 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
                     this.cdr.markForCheck();
                 }
             });
+
+        this.valueChange$.pipe(takeUntil(this.destroy$), debounceTime(100)).subscribe(() => {
+            this.valueChange();
+        });
+
+        if (isPlatformBrowser(this.platformId)) {
+            this.thyClickDispatcher
+                .clicked(0)
+                .pipe(takeUntil(this.destroy$))
+                .subscribe(event => {
+                    if (
+                        !this.elementRef.nativeElement.contains(event.target) &&
+                        !this.menu?.nativeElement.contains(event.target as Node) &&
+                        this.menuVisible
+                    ) {
+                        this.ngZone.run(() => {
+                            this.closeMenu();
+                            this.cdr.markForCheck();
+                        });
+                    }
+                });
+        }
     }
 
     private initSelectionModel() {
@@ -479,15 +514,27 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
 
     private updatePrevSelectedOptions(option: ThyCascaderOption, isActivateInit: boolean, index?: number) {
         if (isActivateInit) {
-            set(option, 'selected', true);
-            this.prevSelectedOptions.push(option);
-        } else {
-            const isSelected = !this.isSelectedOption(option, index);
-            while (this.prevSelectedOptions.length && !this.thyMultiple) {
-                set(this.prevSelectedOptions.pop(), 'selected', false);
+            if (this.thyIsOnlySelectLeaf && option.isLeaf) {
+                set(option, 'selected', true);
             }
-            set(option, 'selected', isSelected);
-            this.prevSelectedOptions.push(option);
+            this.prevSelectedOptions.add(option);
+        } else {
+            if (!this.thyMultiple) {
+                const prevSelectedOptions = Array.from(this.prevSelectedOptions);
+                while (prevSelectedOptions.length) {
+                    set(prevSelectedOptions.pop(), 'selected', false);
+                }
+                this.prevSelectedOptions = new Set([]);
+            }
+            if (this.thyIsOnlySelectLeaf && !option.isLeaf && this.thyMultiple) {
+                set(option, 'selected', this.isSelectedOption(option, index));
+            } else {
+                set(option, 'selected', !this.isSelectedOption(option, index));
+            }
+            if (this.thyIsOnlySelectLeaf && this.thyMultiple && option.parent) {
+                this.updatePrevSelectedOptions(option.parent, false, index - 1);
+            }
+            this.prevSelectedOptions.add(option);
         }
     }
 
@@ -536,7 +583,20 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
                     set(opt, 'selected', true);
                 }
             });
+            this.addParentSelectedState(selectOptions);
         }
+    }
+
+    addParentSelectedState(selectOptions: ThyCascaderOption[]) {
+        selectOptions.forEach(opt => {
+            if (opt.children && opt.children.length && opt.children.every(i => i.selected)) {
+                opt.selected = true;
+                set(opt, 'selected', true);
+                if (opt.parent) {
+                    this.addParentSelectedState([opt.parent]);
+                }
+            }
+        });
     }
 
     setDisabledState(isDisabled: boolean): void {
@@ -564,7 +624,7 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
     }
 
     public isActivatedOption(option: ThyCascaderOption, index: number): boolean {
-        if (!this.isMultiple) {
+        if (!this.isMultiple || this.thyIsOnlySelectLeaf) {
             const activeOpt = this.activatedOptions[index];
             return activeOpt === option;
         } else {
@@ -581,10 +641,19 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
         }
     }
 
+    public isHalfSelectedOption(option: ThyCascaderOption, index: number): boolean {
+        if (!option.selected && this.thyIsOnlySelectLeaf && !option.isLeaf && !this.checkSelectedStatus(option, false)) {
+            return true;
+        }
+        return false;
+    }
+
     public isSelectedOption(option: ThyCascaderOption, index: number): boolean {
         if (this.thyIsOnlySelectLeaf) {
             if (option.isLeaf) {
                 return option.selected;
+            } else {
+                return this.checkSelectedStatus(option, true);
             }
         } else {
             const selectedOpts = this.selectionModel.selected;
@@ -782,8 +851,80 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
         if (option && option.disabled && !this.isMultiple) {
             return;
         }
-        const isSelect = event instanceof Event ? (!this.isMultiple && option.isLeaf ? true : false) : true;
-        this.setActiveOption(option, index, isSelect);
+        const isSelect = event instanceof Event ? !this.isMultiple && option.isLeaf : true;
+        if (this.isMultiple && !option.isLeaf && this.thyIsOnlySelectLeaf && isSelect) {
+            this.toggleAllChildren(option, index, event as boolean);
+        } else {
+            this.setActiveOption(option, index, isSelect);
+        }
+    }
+
+    private toggleAllChildren(option: ThyCascaderOption, index: number, selected: boolean): void {
+        const allLeafs: {
+            option: ThyCascaderOption;
+            index: number;
+        }[] = this.getAllLeafs(option, index, selected);
+        option.selected = selected;
+        while (allLeafs.length) {
+            const { option, index } = allLeafs.shift();
+            option.selected = !selected;
+            this.setActiveOption(option, index, true);
+        }
+        for (let i = 0; i < this.activatedOptions.length; i++) {
+            const option = this.activatedOptions[i];
+            if (isArray(option.children) && option.children.length) {
+                this.setColumnData(option.children, i + 1);
+            }
+        }
+    }
+
+    private getAllLeafs(
+        option: ThyCascaderOption,
+        index: number,
+        selected: boolean
+    ): {
+        option: ThyCascaderOption;
+        index: number;
+    }[] {
+        let allLeafs: {
+            option: ThyCascaderOption;
+            index: number;
+        }[] = [];
+        if (option.children.length > 0) {
+            for (const childOption of option.children) {
+                childOption.parent = option;
+                if (childOption.isLeaf && !childOption.selected === selected) {
+                    allLeafs.push({
+                        option: childOption,
+                        index: index + 1
+                    });
+                } else if (!childOption.isLeaf) {
+                    allLeafs = allLeafs.concat(this.getAllLeafs(childOption, index + 1, selected));
+                }
+            }
+        }
+        return allLeafs;
+    }
+
+    /**
+     * 检查所有所有子项的选择状态, 有一个不符合预期，就直接返回 false
+     * @param option
+     * @param trueOrFalse
+     * @private
+     */
+    private checkSelectedStatus(option: ThyCascaderOption, isSelected: boolean): boolean {
+        if (option.isLeaf) {
+            return option.selected === isSelected;
+        }
+        for (const childOption of option.children) {
+            if (isArray(childOption.children) && childOption.children.length && !this.checkSelectedStatus(childOption, isSelected)) {
+                return false;
+            }
+            if (!childOption.selected === isSelected) {
+                return false;
+            }
+        }
+        return true;
     }
 
     public mouseoverOption(option: ThyCascaderOption, index: number, event: Event): void {
@@ -839,7 +980,7 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
         this.activatedOptions[index] = option;
         for (let i = index - 1; i >= 0; i--) {
             const originOption = this.activatedOptions[i + 1]?.parent;
-            if (!this.activatedOptions[i] || originOption?._id !== this.activatedOptions[i]._id) {
+            if (!this.activatedOptions[i] || originOption?.[this.thyValueProperty] !== this.activatedOptions[i]?.[this.thyValueProperty]) {
                 this.activatedOptions[i] = originOption ?? this.activatedOptions[i];
             }
         }
@@ -882,7 +1023,7 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
                 });
                 this.selectionModel.deselect(currentItem);
             }
-            this.valueChange();
+            this.valueChange$.next();
         }
         if ((option.isLeaf || !this.thyIsOnlySelectLeaf) && !this.thyMultiple) {
             this.setMenuVisible(false);
@@ -903,7 +1044,7 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
         if (isArray(updatedSelectedItems) && updatedSelectedItems.length) {
             this.selectedOptions = updatedSelectedItems[updatedSelectedItems.length - 1].thyRawValue.value;
         }
-        this.valueChange();
+        this.valueChange$.next();
     }
 
     private deselectOption(option: SelectOptionBase) {
@@ -951,7 +1092,7 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
         this.activatedOptions = [];
         this.deselectAllSelected();
         this.setMenuVisible(false);
-        this.valueChange();
+        this.valueChange$.next();
     }
 
     private deselectAllSelected() {
@@ -1004,7 +1145,14 @@ export class ThyCascaderComponent extends TabIndexDisabledControlValueAccessorMi
         return values;
     }
 
-    constructor(private cdr: ChangeDetectorRef, private viewPortRuler: ViewportRuler, public elementRef: ElementRef) {
+    constructor(
+        @Inject(PLATFORM_ID) private platformId: string,
+        private cdr: ChangeDetectorRef,
+        private viewPortRuler: ViewportRuler,
+        public elementRef: ElementRef,
+        private thyClickDispatcher: ThyClickDispatcher,
+        private ngZone: NgZone
+    ) {
         super();
     }