+++ src/cascader/test/cascader.spec.ts
@@ -113,8 +113,23 @@ const multipleOptions = [
                 label: 'liaoyuanshi',
                 children: [
                     {
-                        value: 'kongshanqu',
-                        label: 'kongshanqu',
+                        value: 'longshanqu',
+                        label: 'longshanqu',
+                        isLeaf: true
+                    },
+                    {
+                        value: 'xianqu',
+                        label: 'xianqu',
+                        isLeaf: true
+                    },
+                    {
+                        value: 'dongfengxian',
+                        label: 'dongfengxian',
+                        isLeaf: true
+                    },
+                    {
+                        value: 'dongliaoxian',
+                        label: 'dongliaoxian',
                         isLeaf: true
                     }
                 ]
@@ -525,8 +540,6 @@ describe('thy-cascader', () => {
             dispatchFakeEvent(debugElement.queryAll(By.css(`ul li`))[1].nativeElement, 'mouseover', true);
             dispatchFakeEvent(debugElement.queryAll(By.css(`ul li`))[1].nativeElement, 'click', true);
             fixture.detectChanges();
-            dispatchFakeEvent(document.querySelector('.cdk-overlay-backdrop'), 'click', true);
-            fixture.detectChanges();
         });
 
         it('should select one when click radio and isOnlySelectLeaf is false', done => {
@@ -543,10 +556,6 @@ describe('thy-cascader', () => {
             });
             console.log(debugElement.query(By.css('.form-check-input')).nativeElement);
             debugElement.query(By.css('label')).nativeElement.click();
-
-            fixture.detectChanges();
-            dispatchFakeEvent(document.querySelector('.cdk-overlay-backdrop'), 'click', true);
-            fixture.detectChanges();
         });
 
         it('should menu mouse leave(hover)', () => {
@@ -1023,7 +1032,7 @@ describe('thy-cascader', () => {
 
         it('should show multiple selected label', async () => {
             await fixture.whenStable();
-            const labels = debugElement.queryAll(By.css('.choice'));
+            const labels = debugElement.queryAll(By.css('.choice-item'));
             const selectedValue = component.multipleVal;
             expect(labels.length).toBe(selectedValue.length);
         });
@@ -1060,19 +1069,61 @@ describe('thy-cascader', () => {
 
             expect(component.selectSpy).toHaveBeenCalled();
             expect(component.multipleVal.length).toBe(originSelectedCount + 1);
-            const labels = debugElement.queryAll(By.css('.choice'));
+            const labels = debugElement.queryAll(By.css('.choice-item'));
+            expect(labels.length).toBe(component.multipleVal.length);
+        });
+
+        it('should selection all leafs when click parent node', async () => {
+            component.multipleVal = [
+                ['beijing', 'shixiaqu', 'haidianqu'],
+                ['tianjinshi', 'shixiaqu', 'hepingqu'],
+                ['jilinsheng', 'liaoyuanshi', 'longshanqu']
+            ];
+            fixture.detectChanges();
+            await fixture.whenStable();
+            const originSelectedCount = component.multipleVal?.length;
+            dispatchFakeEvent(debugElement.query(By.css('.form-control')).nativeElement, 'click', true);
+            fixture.detectChanges();
+
+            const firstLevelItem = getOptionByLevel();
+            dispatchFakeEvent(firstLevelItem[2].nativeElement, 'click');
+            fixture.detectChanges();
+
+            const sectionLevelItem = getOptionByLevel(1)[0];
+            dispatchFakeEvent(sectionLevelItem.nativeElement, 'click');
+            sectionLevelItem.query(By.css('label')).nativeElement.click();
+            fixture.detectChanges();
+            await fixture.whenStable();
+
+            expect(component.selectSpy).toHaveBeenCalled();
+            expect(component.multipleVal.length).toBe(originSelectedCount + 3);
+            const labels = debugElement.queryAll(By.css('.choice-item'));
             expect(labels.length).toBe(component.multipleVal.length);
         });
 
         it('should remove item when click x', async () => {
             await fixture.whenStable();
             const originSelectedCount = component.multipleVal?.length;
+            dispatchFakeEvent(debugElement.query(By.css('.form-control')).nativeElement, 'click', true);
+            fixture.detectChanges();
+
             dispatchFakeEvent(debugElement.query(By.css('.thy-icon-close')).nativeElement, 'click', true);
             fixture.detectChanges();
+            await fixture.whenStable();
             expect(component.multipleVal.length).toBe(originSelectedCount - 1);
-            const labels = debugElement.queryAll(By.css('.choice'));
+            const labels = debugElement.queryAll(By.css('.choice-item'));
             expect(labels.length).toBe(component.multipleVal.length);
         });
+
+        it('should close menu when click document', fakeAsync(() => {
+            dispatchFakeEvent(debugElement.query(By.css('.form-control')).nativeElement, 'click', true);
+            fixture.detectChanges();
+            document.body.click();
+            fixture.detectChanges();
+
+            const el = debugElement.query(By.css(`.thy-cascader-picker-open`));
+            expect(el).toBeFalsy();
+        }));
 
         it('should clear item when click clear btn', async () => {
             await fixture.whenStable();
@@ -1094,7 +1145,7 @@ describe('thy-cascader', () => {
             await fixture.whenStable();
 
             expect(component.multipleVal.length).toBe(originSelectedCount - 1);
-            const labels = debugElement.queryAll(By.css('.choice'));
+            const labels = debugElement.queryAll(By.css('.choice-item'));
             expect(labels.length).toBe(component.multipleVal.length);
 
             const updateFirstLevelItem = getOptionByLevel();
@@ -1104,7 +1155,7 @@ describe('thy-cascader', () => {
         it('should show nothing when ngModel is []', fakeAsync(() => {
             component.multipleVal = [];
             fixture.detectChanges();
-            const labels = debugElement.queryAll(By.css('.choice'));
+            const labels = debugElement.queryAll(By.css('.choice-item'));
             expect(labels.length).toBe(0);
         }));
 
@@ -1113,7 +1164,7 @@ describe('thy-cascader', () => {
             fixture.detectChanges();
             await fixture.whenStable();
             const originSelectedCount = component.multipleVal?.length;
-            const labels = debugElement.queryAll(By.css('.choice'));
+            const labels = debugElement.queryAll(By.css('.choice-item'));
             expect(labels.length).toBe(originSelectedCount);
         });